@@ -24,43 +24,40 @@
 	},
 	"homepage": "https://github.com/Steff-ee/Steff-ee.github.io#readme",
 	"devDependencies": {
-		"@babel/core": "^7.9.0",
-		"@babel/plugin-proposal-class-properties": "^7.8.3",
-		"@babel/plugin-proposal-object-rest-spread": "^7.9.0",
-		"@babel/plugin-transform-modules-commonjs": "^7.9.0",
-		"@babel/preset-env": "^7.9.0",
-		"@babel/preset-react": "^7.9.1",
-		"@babel/preset-typescript": "^7.9.0",
-		"@types/enzyme": "3.10.3",
-		"@types/history": "4.7.3",
-		"@types/jest": "24.0.18",
-		"@types/lodash": "4.14.137",
-		"@types/node": "12.7.2",
+		"@babel/core": "^7.18.6",
+		"@babel/plugin-proposal-class-properties": "^7.18.6",
+		"@babel/plugin-proposal-object-rest-spread": "^7.18.6",
+		"@babel/plugin-transform-modules-commonjs": "^7.18.6",
+		"@babel/preset-env": "^7.18.6",
+		"@babel/preset-react": "^7.18.6",
+		"@babel/preset-typescript": "^7.18.6",
+		"@types/enzyme": "3.10.12",
+		"@types/history": "4.7.11",
+		"@types/jest": "28.1.4",
+		"@types/lodash": "4.14.182",
+		"@types/node": "18.0.1",
 		"@types/react": "^18.0.14",
-		"@types/react-dom": "^16.9.5",
-		"@types/react-router": "^5.1.4",
-		"@types/react-router-dom": "^5.1.3",
-		"@types/webpack-env": "^1.15.1",
+		"@types/react-dom": "^18.0.6",
+		"@types/react-router": "^5.1.18",
+		"@types/react-router-dom": "^5.3.3",
+		"@types/webpack-env": "^1.17.0",
 		"babel-core": "^6.26.3",
-		"babel-loader": "^8.0.6",
-		"cross-env": "^7.0.2",
-		"css-loader": "^3.4.2",
-		"file-loader": "^6.0.0",
-		"html-webpack-plugin": "^4.0.0",
-		"node-sass": "^4.13.1",
-		"react-hot-loader": "^4.12.20",
+		"babel-loader": "^8.2.5",
+		"cross-env": "^7.0.3",
+		"css-loader": "^6.7.1",
+		"eslint": "^8.19.0",
+		"file-loader": "^6.2.0",
+		"html-webpack-plugin": "^5.5.0",
+		"node-sass": "^7.0.1",
 		"react-use": "^17.4.0",
-		"sass-loader": "^8.0.2",
-		"source-map-loader": "^0.2.4",
-		"style-loader": "^1.1.3",
-		"tslint": "^6.1.0",
-		"tslint-config-prettier": "^1.18.0",
-		"tslint-react": "^4.2.0",
+		"sass-loader": "^13.0.2",
+		"source-map-loader": "^4.0.0",
+		"style-loader": "^3.3.1",
 		"typescript": "^4.7.4",
-		"url-loader": "^4.0.0",
-		"webpack": "^4.42.0",
-		"webpack-cli": "^3.3.11",
-		"webpack-dev-server": "^3.11.2"
+		"url-loader": "^4.1.1",
+		"webpack": "^5.73.0",
+		"webpack-cli": "^4.10.0",
+		"webpack-dev-server": "^4.9.3"
 	},
 	"dependencies": {
 		"@emotion/react": "^11.9.3",
@@ -68,21 +65,22 @@
 		"@fortawesome/fontawesome-svg-core": "^1.2.26",
 		"@fortawesome/free-solid-svg-icons": "^5.12.0",
 		"@fortawesome/react-fontawesome": "^0.1.8",
-		"@mui/material": "^5.8.5",
-		"d3-ease": "^1.0.6",
-		"enzyme": "3.10.0",
-		"enzyme-to-json": "3.4.0",
-		"history": "4.9.0",
-		"immutable": "4.0.0-rc.12",
+		"@mui/material": "^5.8.7",
+		"d3-ease": "^3.0.1",
+		"enzyme": "3.11.0",
+		"enzyme-to-json": "3.6.2",
+		"history": "5.3.0",
+		"immutable": "4.1.0",
 		"lodash": "^4.17.21",
 		"react": "^18.2.0",
 		"react-dom": "^18.2.0",
 		"react-media": "^1.10.0",
-		"react-particles-js": "^2.7.0",
-		"react-progressive-loader": "^0.4.1",
-		"react-router": "^5.1.2",
-		"react-router-dom": "^5.1.2",
-		"react-spring": "^9.0.0-canary.809.5.f01ecc2",
-		"sunrise-sunset-js": "^2.2.1"
+		"react-progressive-loader": "^0.4.3",
+		"react-router": "^6.3.0",
+		"react-router-dom": "^6.3.0",
+		"react-spring": "^9.4.5",
+		"react-tsparticles": "^2.1.3",
+		"sunrise-sunset-js": "^2.2.1",
+		"tsparticles": "^2.1.3"
 	}
 }
