@@ -1,5 +1,5 @@
 import React, { useState } from 'react'
-import { HashRouter, Redirect, Route, Switch } from 'react-router-dom'
+import { HashRouter, Navigate, Route, Routes } from 'react-router-dom'
 import { MediaProvider } from '../components/mediaProvider'
 import { ClassicPageContainer } from '../modes/classic/classicPageContainer'
 import { SeasonsProvider } from '../modes/seasons/seasons'
@@ -7,13 +7,11 @@ import '../scss/root.scss'
 import { PageRoutes, RouteContext } from '../shared/helpers/routes'
 import { OpenPostsProvider } from '../shared/posts/openPosts'
 import { PivotRoutes } from '../shared/posts/post.types'
-import { IsNavBarOpenContext } from '../shared/presentational/components/navBarHelpers'
 import { ColorsProvider } from '../shared/presentational/hooks/useColors'
 import { CatsPage } from './cats/cats'
 import { PageNotFound } from './pageNotFound'
 
 export const Root: React.FunctionComponent = (): JSX.Element => {
-	const [isNavBarOpen, setIsNavBarOpen] = useState<boolean>(false)
 	const [prevPivots, setPrevPivots] = useState<{ [page: string]: PivotRoutes | undefined }>({})
 	const setPrevPivot = (page: string, pivot: PivotRoutes): void => {
 		setPrevPivots({ ...prevPivots, [page]: pivot })
@@ -24,28 +22,28 @@ export const Root: React.FunctionComponent = (): JSX.Element => {
 			<OpenPostsProvider>
 				<SeasonsProvider>
 					<ColorsProvider>
-						<IsNavBarOpenContext.Provider value={{ isNavBarOpen, setIsNavBarOpen }}>
-							<link
-								href="https://fonts.googleapis.com/css?family=Source+Code+Pro:300,400|Open+Sans:400,400i,600|Montserrat:400,500,600&display=swap"
-								rel="stylesheet"
-							/>
-							<MediaProvider>
-								<HashRouter>
-									<Switch>
-										<Redirect exact={true} from="/" to={PageRoutes.Home} />
-										<Route path={`/home`} component={ClassicPageContainer} />
-										<Route path={`/stories`} component={ClassicPageContainer} />
-										<Route path={`/games`} component={ClassicPageContainer} />
-										<Route
-											path={`/conjecture`}
-											component={ClassicPageContainer}
-										/>
-										<Route path={`/cats`} component={CatsPage} />
-										<Route component={PageNotFound} />
-									</Switch>
-								</HashRouter>
-							</MediaProvider>
-						</IsNavBarOpenContext.Provider>
+						<link
+							href="https://fonts.googleapis.com/css?family=Source+Code+Pro:300,400|Open+Sans:400,400i,600|Montserrat:400,500,600&display=swap"
+							rel="stylesheet"
+						/>
+						<MediaProvider>
+							<HashRouter>
+								<Routes>
+									<Route path={`/`} element={<Navigate to={PageRoutes.Home} />} />
+									<Route path={`/cats`} element={<CatsPage />} />
+									<Route
+										path={`/:page/:pivot/:postId`}
+										element={<ClassicPageContainer />}
+									/>
+									<Route
+										path={`/:page/:pivot`}
+										element={<ClassicPageContainer />}
+									/>
+									<Route path={`/:page`} element={<ClassicPageContainer />} />
+									<Route element={<PageNotFound />} />
+								</Routes>
+							</HashRouter>
+						</MediaProvider>
 					</ColorsProvider>
 				</SeasonsProvider>
 			</OpenPostsProvider>
