@@ -1,8 +1,7 @@
-import React, { useContext } from 'react'
-import Particles from 'react-particles-js'
+import React from 'react'
+import Particles from 'react-tsparticles'
 import { Colors } from '../../helpers/constants'
 import { CircadianMood, getCircadianMood } from '../../helpers/time'
-import { IsNavBarOpenContext } from './navBarHelpers'
 
 export interface IFirefliesProps {
 	style?: React.CSSProperties
@@ -19,7 +18,6 @@ interface IHighlightProps {
 
 export const Fireflies: React.FunctionComponent<IFirefliesProps> = (props) => {
 	const { style, width, height } = props
-	const { isNavBarOpen } = useContext(IsNavBarOpenContext)
 	const mood = getCircadianMood()
 	const animSpeed = 3
 	const moveSpeed = 2
@@ -27,7 +25,7 @@ export const Fireflies: React.FunctionComponent<IFirefliesProps> = (props) => {
 	const bubbleSize = 6
 
 	let count
-	if (mood === CircadianMood.Night || isNavBarOpen) {
+	if (mood === CircadianMood.Night) {
 		count = 60
 	} else if (mood === CircadianMood.Day) {
 		count = 0
@@ -35,29 +33,26 @@ export const Fireflies: React.FunctionComponent<IFirefliesProps> = (props) => {
 		count = 20
 	}
 
-	let options: IHighlightProps
-	if (isNavBarOpen) {
-		options = {
-			maxSize: 5,
-			minSize: 2,
-			maxOpacity: 1,
-			minOpacity: 0.8,
-		}
-	} else {
-		options = {
-			maxSize: 3,
-			minSize: 1,
-			maxOpacity: 0.8,
-			minOpacity: 0.2,
-		}
+	const options: IHighlightProps = {
+		maxSize: 5,
+		minSize: 2,
+		maxOpacity: 1,
+		minOpacity: 0.8,
+	}
+	// options = {
+	// 	maxSize: 3,
+	// 	minSize: 1,
+	// 	maxOpacity: 0.8,
+	// 	minOpacity: 0.2,
+	// }
+	const particlesLoaded = async () => {
+		console.log('loaded!')
 	}
 
 	return (
 		<Particles
-			style={style}
-			width={width}
-			height={height}
 			params={{
+				fullScreen: { enable: true },
 				particles: {
 					number: {
 						value: count,
